@@ -1,5 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using DevExpress.Data.Filtering;
+using DevExpress.Xpo;
+using Microsoft.AspNetCore.Mvc;
 using SimpleJwt;
+using Solution1.Module;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -8,7 +11,9 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using WebApiDemo.Controllers.Helpers;
 using Xpo.RestDataStore;
+
 
 namespace WebApiDemo.Controllers
 {
@@ -17,28 +22,67 @@ namespace WebApiDemo.Controllers
     public class LoginController : ControllerBase
     {
         public const string Key = "JIHHKS304SS1E9SXT2P1OXHN3Y1XHETRVSVB71EPYHR8U6FO2GMKYX4WB3NW0GTZ9BFWORZ1XACHGI26SM5B2J27X89DOSYSY18S0TAKRUWYVXQ2BRDNLR70JI8PRVNE";
-        public const string Issuer = "Jose Manuel Ojeda";
+        public const string Issuer = "Jose Columbie";
 
         [HttpPost]
         [Route("[action]")]
         public async Task<LoginResult> Login()
         {
             //List of standard Payload claims https://en.wikipedia.org/wiki/JSON_Web_Token#Standard_fields
-            var ParametersByte = await Request.GetRawBodyBytesAsync();
-            var LoginParameters = RestApiDataStore.GetObjectsFromByteArray<LoginParameters>(ParametersByte);
-            if (LoginParameters.Username == "Joche" && LoginParameters.Password == "123")
+
+            byte[] ParametersByte = await Request.GetRawBodyBytesAsync();
+
+            LoginParameters LoginParameters = RestApiDataStore.GetObjectsFromByteArray<LoginParameters>(ParametersByte);
+            string Database;
+            string Server;
+            Employee User = null;
+            try
             {
-                JwtPayload InitialPayload;
-                InitialPayload = new JwtPayload {
-                { JwtRegisteredClaimNames.NameId, LoginParameters.Username },
-                { JwtRegisteredClaimNames.Iat, JwtHelper.ConvertToUnixTime(DateTime.Now).ToString() },
-                  { JwtRegisteredClaimNames.Iss, Issuer },};
+
+                string UserName = WebUtility.UrlDecode(LoginParameters.Username);
+                string Password = WebUtility.UrlDecode(LoginParameters.Password);
+               
+                Server = WebUtility.UrlDecode(LoginParameters.Server);
+
+                Database = WebUtility.UrlDecode(LoginParameters.Database);
+                UnitOfWork UoW = XpoProxyHelper.GetUnitOfWork(Database, Server);
+
+                User = UoW.FindObject<Employee>(new BinaryOperator("UserName", UserName));
+               
+              
+                if (User == null)
+                {
+                    return new LoginResult() { Authenticated = false, Token = "" };
+                }
+                if (!User.ComparePassword(Password))
+                {
+
+                    return new LoginResult() { Authenticated = false, Token = "" }; //TODO invalid password
+
+                }
+
+               
+
+            }
+            catch (Exception exception)
+            {
+                return new LoginResult() { Authenticated = false, Token = "", ErrorMessage= exception.Message };
+            }
+      
+
+             JwtPayload InitialPayload;
+                InitialPayload = new JwtPayload
+                {
+                    { JwtRegisteredClaimNames.NameId, LoginParameters.Username },
+                    { JwtRegisteredClaimNames.Iat, JwtHelper.ConvertToUnixTime(DateTime.Now).ToString() },
+                    { JwtRegisteredClaimNames.Iss, Issuer },
+                    { "DatabaseId", Database },
+                    { "ServerId", Server },
+                };
 
                 var StringToken = JwtHelper.GenerateToken(Key, InitialPayload);
-                return new LoginResult() { Authenticated = true, Token = StringToken };
-            }
-            else
-                return new LoginResult() { Authenticated = false, Token = "" };
+                return new LoginResult() { Authenticated = true, Token = StringToken, CurrentUserId = User?.Oid.ToString()};
+           
         }
     }
 }
