@@ -9,11 +9,16 @@ namespace Xpo.RestDataStore
     {
         public bool Authenticated { get; set; }
         public string Token { get; set; }
+        public string CurrentUserId { get; set; }
+        public string ErrorMessage { get; set; }
 
-        public LoginResult(bool authenticated, string token)
+
+        public LoginResult(bool authenticated, string token, string userid, string error = "")
         {
             Authenticated = authenticated;
             Token = token;
+            CurrentUserId = userid;
+            ErrorMessage = error;
         }
 
         public LoginResult()
